Move the coroutine standard library out of the public hhi
Summary:
Public hhi files are type checked and slurped by HHVM. Neither our type checker nor runtime understand `coroutine` or `suspend` as syntaxes. So I have to move the coroutine standard library out of this public space and into somewhere private. In the future, the coroutine standard library will become part of the public HHI space. But we're not there yet.

I've updated the tests to pull in //both// the public HHI library and the coroutine standard library where appropriate.

Reviewed By: ericlippert

Differential Revision: D5298283

fbshipit-source-id: 48452f16093f2a34d639c1aaa0e7c26ad93e70ff