Return the coroutine unit singleton instead of the intermediate, suspended result
Summary:
When suspending, for type checking purposes, we should return the `CoroutineUnit`, rather than the intermediate, suspended, result. The intermediate suspensions may be of disparate types. However, the state machine
exposes a public interface that says it will return a `CoroutineResult` whose type parameter is the type of the coroutine. It is a type checker error to return a suspended `CoroutineResult<Foo>` when the coroutine declares its return type to be `CoroutineResult<Bar>`. So instead we just return the singleton `CoroutineUnit` instance.

Reviewed By: ericlippert

Differential Revision: D5278125

fbshipit-source-id: eb20e3e901fb508b6a9a92eb44c1add0e1681ef2