emit return-type + parameters in autocomplete
Summary:
LSP doesn't have any way to specify "return-type" or "parameters" for CompletionItem. But Atom and AutocompletePlus like to display those in the left and right columns.

This diff invents two new fields in the LSP protocol. We will propose them to the LSP project on due turn, once we've verified that they work out okay. It sends them over the wire in response to textDocument/completion

The diff also makes one related change. The 'detail' field is meant to show the signature of the thing. It was stripping parentheses so that function types like `(function(int): string)` wouldn't have the outer parentheses. But it was also stripping parentheses in tuple types like `(string, int)`. This diff makes it only strip in the case of function types.

Reviewed By: arxanas

Differential Revision: D5324300

fbshipit-source-id: 9445d4e859fcf33b4d68ebbe47f6c1509d4bc582