*easy* autocomplete emits parentheses+args for method calls
Summary:
The previous persistent-connection followed the behavior of Eclipse+IntelliJ by autocompleting `$c->|` into `$c->foo($arg1)`. This diff does the same to LSP.

Note: the previous persistent-connection would auto-insert parentheses+args if and only if there was no open-parenthesis on the rest of the line. This diff instead does it if and only if there's whitespace where the caret is. It felt nicer this way.

Reviewed By: arxanas

Differential Revision: D5329584

fbshipit-source-id: b60829f3ac166291d7e66126875f507dcbba9f26