strengthen return type of Shapes::idx
Summary: When a field is required to exist in a shape (i.e. it is neither optional nor nullable), Shapes::idx's return-type should be the type of the field instead of `Toption`. This will also emit a lint warning that suggests using a subscript expression instead. This feature is guarded by `experimental_stronger_shape_idx_ret`.

Reviewed By: andrewjkennedy

Differential Revision: D4931888

fbshipit-source-id: a4313e5e2b1c509a6aea667fd473e6af0905f42d