basic genetic programming script for executing fuzzer
Summary: This script takes as an input a hhas file, and will apply mutations to it for a user-specified number of generations. At the end of each generation, files that fail more than a user-specified number of verification tests will be culled, and the remaining files will be used as new inputs for the next generation of the script. The user can also specify a number of threads for parallel execution.

Reviewed By: paulbiss

Differential Revision: D5281345

fbshipit-source-id: a2da98949f7b14c684029822178ac2f7caa4d079