Standardizing allowed use of 'abstract' keyword in classish declarations
Summary:
Prior to this diff, both HHVM and the original parser would produce errors on "abstract interface" and "abstract trait" declarations, but the full fidelity parser would not. In this diff, I've added errors 2042 and 2043 to the full fidelity parser to ensure that it also produces errors on those declarations. The updated behavior of the full fidelity parser is demonstrated in test_abstract_classish_errors.php. (HHVM and the original parser also produce errors only on lines 6 and 18. However, both of them mark "interface" and "trait" as the location of the error, giving messages "Fatal error: unexpected (T_INTERFACE/T_TRAIT)", and "Expected 'class' (not '[interface/trait]') ", respectively. I've deviated from this precedent and made the full fidelity parser specify the keyword 'abstract' as the location of the error.)

Also made a minor wording change to error 2041, for clarity.

Differential Revision: D5338352

fbshipit-source-id: 839e2e1d0f88885b5dffabf0e588ec8134356369