Don't create LivePrologues for PGO'd functions while RetranslateAll is in flight
Summary:
This diff fixes a race in which a live prologue could be emitted between when we
reset the func profile prologues and we create and set the optimized ones.  We'd
eventually replace those live prologues with the optimized ones in the func
prologue table.  However, other function calls could still be indefinitely bound
to the live prologues.

Reviewed By: mxw

Differential Revision: D5353452

fbshipit-source-id: 10ed63d09d2f306a3501b625a250f9483a7a3a70