Add a (soft) limit on the number of guards per tracelet
Summary:
The more guards a tracelet has, the bigger the chances that it'll need a
retranslation, and also that the translations are going to be bigger.  Normally,
we break tracelets after a few bytecode instructions because of control flow or
consumption of a value produced by the tracelet with unknown type.  However, in
the lack of any of these, a tracelet may grow really big in some degenerated
cases, potentially requiring large retranslations and bloating the JITed code.
This diff avoids most of these degenerated cases by introducing a limit on the
number of guards per tracelet.  This is a soft limit which, when crossed, will
trigger the end of the tracelet.

Reviewed By: paulbiss, jano

Differential Revision: D5362076

fbshipit-source-id: c3fa95d45187f9909b2c27e78b85bc29cb46c2ab