Emit GenericRetDecRefs for functions with too many locals
Summary:
If a function has too many locals, just guarding all of them can produce a lot
of code.  Furthermore, guarding on too many locals increases the chance of any
of them being polymorphic, which will trigger retranslations to produce another
large retranslation.  To avoid this issue, this diff introduces a limit on the
total number of locals that a function is allowed to have in order for its
locals to be DecRef'd inline.  This new limit is for all locals, even if their
type is not ref-counted.  (We already had a similar limit for ref-counted locals
only.)

Reviewed By: jano, paulbiss

Differential Revision: D5362092

fbshipit-source-id: fade8aee3da4d1fdd8d7823e1d760d570e6e4c7a