Add global type environment to handle constraints that are global to function bodies
Summary:
In order to allow type constants to work under the generic constraint system, we need them to add constraints to the environment during typechecking a function body. These constraints need to distinguish between global constraints, those that persist through local environments, and local constraints, those that are specific to a scope.

https://fb.facebook.com/groups/721152551318888/permalink/931538073613667/ for more info

Reviewed By: dlreeves

Differential Revision: D4406475

fbshipit-source-id: 9fb1bbf8a3dbf46d9ae59b3229b848262abfa1b9