Don't crash or leak in fb_json_decode if we get duplicate keys.
Summary:
When we json_decode a string <= SimpleJsonMaxLength with the JSON_FB_LOOSE flag set, we first try to do it using SimpleParser.
SimpleParser was not checking if MixedArray::MakeMixed returned nullptr (which happens if you give it duplicate keys), and would crash.
Also, the MixedArray::MakeMixed code paths that return nullptr were leaking the MixedArary they were in the process of constructing.

Reviewed By: edwinsmith

Differential Revision: D5367734

fbshipit-source-id: 8c158b51b35b96690c60114d07053b6501edac03