vget _n pattern for iterators
Summary:
Adds special patterns to cope with VGetL on unnamed locals in the context of compiling unnatural foreach statements in PHP. In cases
like foreach ($array as $key => &$$v) we do take a reference to an unnamed local, but it does not escape and the local is unset within a couple
of instructions, so it is safe to treat this as a read and then an unset of the local. This diff adds rather specific and hacky special case
rules to cope with the associated patterns of code.

Reviewed By: hubyrod

Differential Revision: D5337863

fbshipit-source-id: 829bb3c6e467ebfeb2d4fdde770e9b60c130e440