Validation functions for FFP tree
Summary:
Finally, after all the type-juggling, here are the validation functions for the full fidelity parse tree. From the schema defintion, generate functions that check only the expected node types occur in the expected places. When they do, return a parse tree that reflects these invariants in its own type.

Note that `fun x -> validate (invalidate x)` and `fun x -> invalidate (validate x)` should both be `id`.

This diff works mostly as an announcement and a way to chime in. There are some cleanups to be expected in the near future.

Reviewed By: ericlippert

Differential Revision: D5155600

fbshipit-source-id: 455fa68556c309ee4c7a12f1a557eef60bc57572