Fix performance issue with type history for exceptional flow
Summary:
Hack maintains a "history" of the types that a local variable is assigned, because in general an exception can divert control to an `catch` block in which that variable might have any of those types.
Hack creates a union of those types (a `Tunresolved`) for the exception handler itself. For example, in
```
try {
  $x = 1;
  foo();
  $x = 2.3;
  bar();
  $x = 'a';
} catch (Exception $e) {
  // Here $x has type int | float | string
}
```
In a number of places Hack simply adds types to a list, checking only that there are no exact syntactic duplicates of the type.
But any slight variation - e.g. with different `Reason` information, or through the indirection of a type variable - will count as distinct.
In large code blocks this can lead to the build-up of essentially-duplicate types in the "history", and consequent quadratic behaviour.
The fix is simply to use the recently-added `ty_equal` function that doesn't care about `Reason` information, and first expand the type so that type variable indirections are ignored.

Longer term, we could replace the list by a set, now that we have `Typing_set.ml`.

Reviewed By: oulgen

Differential Revision: D5369063

fbshipit-source-id: ceba516d8d307c66624e9b9215058899f4ae0743