fuzzer reorder preserves stack, and can move instruction sequences instead of indivdiual instructions
Summary: The reordering mutation will randomly choose two bytecode sequences with the same stack signature (the stack has the same height at the beginning and end of both, respectively) and swap their places in the code. So far this has been more likely than the other mutations to produce trivially broken code, but I think with the right input file this could probably uncover a couple of interesting bugs.

Reviewed By: nbenton

Differential Revision: D5344097

fbshipit-source-id: aa28f2a82b28a20d4749e02fcc96fb8088272270