PHP debugger: break if eval fails on conditional breakpoint
Summary:
If a breakpoint has a condition set on it, but the condition expression fails to evaluate, the debugger should err on the side of breaking in anyway, rather than missing the user's breakpoint. This is safer behavior, since over actively hitting a breakpoint is at worst annoying, while missing a breakpoint could be serious - like missing a repro condition for a bug, or letting the debugger go down a code path with side effects when the user didn't expect it.

When this happens, the evaluation error is already printed to the error stream (stderr, slog, whatever the webserver is doing) so it appears in the console. The user then knows to fix the breakpoint condition or remove it before continuing

Reviewed By: alexmalyshev

Differential Revision: D5369290

fbshipit-source-id: c384175fc768ca7dcb318aa95e9e44780f98dc42