Adding documentation for predecessor function used in keyword autocompletion
Summary:
Also adding comment in full_fidelity_syntax.ml to note that the children function should return the children in the same order as they appear in a source file. i.e. For a tuple expression `tuple(a,b)`, its children are, in order:
- tuple expression keyword
- tuple expression left paren
- tuple expression items
- tuple expression right paren

Everything already in the file should conform to this rule, since this invariant is also used for the parentage functions for each type of full_fidelity_syntax.

Reviewed By: arxanas

Differential Revision: D5360243

fbshipit-source-id: d7a804430dd3b5e59fae02d958b33521b243133a