Remove wrong code in cGetPropImpl
Summary:
Everything below is in the context of non-repo mode.

When HHBBC is enabled under non-repo mode, it can sometimes determin attributes regarding magic prop methods, e.g. hphp/test/quick/CGetMVar.php. When the prop being accessed has a pre-determined offset, and when HHBBC sets proper attributes after analyzing magic props, it will enter this troubled branch.

In this branch, the `if` block is wonrg. The `UnboxPtr` is being called using `base` which makes no sense since `propAddr` should be the correct address to use. And the code following this `if` block is using `propAddr` properly.

Reviewed By: mxw

Differential Revision: D5372337

fbshipit-source-id: f1828fdd117059a8c307d0dc0bea405ebd78f23a