Accelerate pointer filtering with a per-slab crossing map.
Summary:
The slowest phase in GC is marking the heap, and the bottleneck is
looking up objects with a binary search in PtrMap. Introduce a Slab
header, containing a "crossing map" which allows a faster way to find
the beginning of an object, given an interior pointer.

Marker::init() still builds a PtrMap for the heap, but now it only
contains large objects and slabs, so it's much smaller; pointer
filtering now does a (faster) search in this toplevel map, followed
by a crossing-map lookup for the common case of pointers within slabs.

The crossing map requires 1 byte per "Line" of slab space. Using
a LineSize of 256 bytes (found experimentally), the overhead is 0.4%.
Larger lines save space but require more heap scanning. Similarly,
smaller lines allow yet faster lookups, with more map overhead.

Reviewed By: alexeyt

Differential Revision: D5369231

fbshipit-source-id: 9b39d77d7ec7053e5a8b02c1412fdb9342365a03