Add 'use type' to Hacklang
Summary:
I went for `use type` in the syntax as it's not just for classes.
Even though it's inconsistent, I went for calling it 'class' internally, as
the most consistent thing to do woudl be to call them type aliases, which just
seems like it's going to be confusing. I don't feel strongly about this though.

Reviewed By: jamesjwu

Differential Revision: D5352626

fbshipit-source-id: 2d714b21d6eb9c0889759df5e972c8bd2656c04e