Do not count parsing/emitting towards request's memory limit
Summary:
A giant PHP file, such as an autoload map, can trigger memory limit of a PHP
request during its parsing and emitting. When this happens, HHBC still gets
produced and cached, but the PHP request fails with OOM. The next request
reuses the cached HHBC and does not hit this error.

Since the parsing/emitting work is not done on the behalf of the first request,
but on behalf of all future requests, it should not be counted towards memory
limits. This diff achieves that by suppressing OOM inside PHP file
parser/emitter and GetScalarArrayKey used by UnitEmitter's static array merging
mechanism.

Reviewed By: paulbiss

Differential Revision: D5380050

fbshipit-source-id: cb0bf3184aa3504b46bdc91fc265e6ab890bf642