Fix race in retranslateAll vs hphp_process_exit
Summary:
We could be adding jobs to the dispatcher, even while the dispatcher
is being shut down by hphp_process_exit; this causes
dispatcher().waitEmpty() to return immediately, and the jobs are
destroyed while they're still running.

JobQueueDispatcher::enqueue isn't thread safe, so needs a lock. We
only ever create the dispatcher from one thread, so we don't need a
lock for that.

Reviewed By: paulbiss

Differential Revision: D5375952

fbshipit-source-id: 674bd50890396d7e7fa8f334df727b6eeb7de21e