Coroutine refactoring: Nest destructurings of syntaxes and eliminate unnecessary uses of option
Summary:
At various places in the coroutine lowering, we have the following kind of pattern:

  1. Destructure a Syntax node by calling `syntax`, and matching on the expected syntax type
  2. Extract a child node of the syntax
  3. Pass the child node to another function, and repeat this destructuring process from that child node

However, we can destructure //without// calling the `syntax` function. So that means we can do all of that destructuring in a single `match` statement. This diff implements that in two large sections of the `coroutine_lowerer`.

Additionally, I originally wrote some of the coroutine lowering code to use `option`, and return `None` if it got an unexpected syntax element. Since we only care about working on valid syntax trees, I've converted these `option`s to just be a value or throw an exception.

**N.B.** Once drphil's work with the validated syntax tress is done (looks like it's almost there, in D5155600), these changes will make it easier to directly plug into the validated syntax tree.

Reviewed By: ericlippert

Differential Revision: D5314847

fbshipit-source-id: 813bc40a3e188f2966f09cbe6ee69f45c89c29c0