Fix subtype, couldBe, and equals operators on HHBBC specialized array types
Summary:
The subtype, couldBe, and equals operators on HHBBC specialized array types
didn't deal with all cases properly. Namely, it largely assumed that the packed
and map representations couldn't overlap. They can, however, because there's
nothing stopping a map type from having keys 0 to N-1. Fix these errors and add
comments in the non-obvious places.

The other un-accounted for overlap is MapN with a constant key type. This
implies that the MapN only contains arrays with a single key (because the key is
constant so it can't repeat), which means it could overlap with various Map and
Packed representations. Instead of dealing with this, its better to forbid this
case and assert it as an invariant. If a MapN has a constant key, its better to
represent it as a Map instead. This can only happen in one place right now, so
add the check there and convert to Map.

Finally, remove the `len` parameter from PackedN. The various operators don't
really account for it, and its not used in practice.

Reviewed By: markw65

Differential Revision: D5352110

fbshipit-source-id: fdfbc889cf061717d0126a00a6ea74844e5811a4