Hoist table/mask calculation for findForNewInsert out of loops.
Summary: There are 2 versions of findForNewInsert{,Warn}: the 3 argumnt form where the hash table pointer and hash mask are passed in, and the 1 argument form that calculates the table and mask and calls the 3 argument form. Don't use the 1 argument form in loops: compute table and mask outside of the loop and pass them in.

Reviewed By: edwinsmith

Differential Revision: D5380450

fbshipit-source-id: 0d63669c576c0733ebb9729d64a4a556cd651c10