Start refactoring coroutine lowering for functions, lambdas
Summary:
I'm going to make a series of refactorings to the coroutine lowering to support lowering of (1) functions outside of classes, and (2) lambdas.

Previous to this diff, the lowering code assumed that the coroutine being lowered was in a class, and passed in the class node. All we need from it is the name of the class (to generate the closure name, since the closure class cannot be nested in Hack) and the generic parameters (to copy to the closure class.) We now just pass those in; we can pass in a dummy name and an empty parameter list for coroutines outside of classes.

Reviewed By: michaeltingley

Differential Revision: D5343074

fbshipit-source-id: 4b8c7e98f05beca5b4b9cf7b0d5a852155c5cd61