Fix a couple of store-elim bugs
Summary:
Store-elim can't ignore AssertLoc or AssertStk because eliminating a
store, or sinking it past them, could affect the value they're
supposedly asserting.

Generally this isn't an issue because load-elim will convert them to
AssertType (or eliminate them altogether) - but for cases where it
fails, we should disable store-elim.

While doing this, I noticed a potential problem with the self-store
optimization: when it finds a pure load from a self-store candidate,
it doesn't check that the load really is the source of the store (eg
there could be an earlier load too; and the earlier load may not be
available).

Reviewed By: ricklavoie

Differential Revision: D5384861

fbshipit-source-id: 309a086645e1b2acf3d492f11e27adc696000db0