Refactor function lowering to use parented aggregating rewrite
Summary:
The current top-level algorithm for rewriting coroutines is a bit overcomplicated as a result of the early, now rejected decision to generate helper methods rather than lambdas for the rewritten coroutine body in the case where the coroutine is a method of a class.

Now that we are rewriting bodies in place and generating a closure class, we can simplify the rewriting algorithm overall to:

* recursively rewrite the bodies of all the lambdas, anonymous functions, methods and functions into constructions of closure classes with a lambda as the argument
* accumulate closure class implementations as we go
* stick the closure classes at the end of the file

I'm going to gradually refactor the existing implementation to do this. In this diff, we now rewrite top-level coroutine functions *first*, accumulate their closures, then do the rest of the rewriting, and finally append the new closures.

Eventually we'll get it so that all the rewriting happens like that, and much of the logic for dealing with classes will just go away.

The only change in the output will be that the closure classes will be generated at the end of the file rather than next to the class or function which caused its creation.

Reviewed By: michaeltingley

Differential Revision: D5352553

fbshipit-source-id: 9b5ea2754566bf25209c67c68627eee5cc65239d