Add LSP testing script
Summary:
Add this LSP testing script that I use to test the language server. This doesn't in itself create any tests or test frameworks. It might be useful to make a testing framework, though.

alexchow commented this on the topic of testing in Python versus OCaml:

> * depends what your'e doing. Ocaml libraries are mostly lacking, so sometimes it's easier to do things in Python.
> * if it's not an integration test involving doing mercurial commands on disk, i'd suggest avoiding python.
> * if it's an integration test, see if test_harness.ml provides what you need.
> * also, check out integration_ml which has an artificial run loop
> * and also entirely ocaml-land.

So we may want to do it in OCaml. (In particular, we do actually have JSON libraries for OCaml.)

Note that this script is set up for interactive use, and currently doesn't terminate the connection to the language server in a graceful way. Instead, it just keeps stdout open, which means that the process will hang until you kill it.

Differential Revision: D5391324

fbshipit-source-id: 721ac86eee1506ee6b3bcf926b5338b4959392e3