Create data type for coroutine context to generate closure signature
Summary:
We tramp around a lot of data that we need to generate the coroutine closure class signature; in an earlier diff Michael suggested that this be represented by its own type.

This idea has great merit; in this diff, just the declaration of the type. In subsequent diffs I'll plumb it through and get rid of some of the tramp parameters

Reviewed By: michaeltingley

Differential Revision: D5360293

fbshipit-source-id: 8d4037e41b6f26c1f3c5200437fcd1aef0e19295