Catch improper OODeclExists opcodes with runtime errors
Summary: HHVM was crashing on OODeclExists opcodes where the top two elements on the stack were not the right type (bool and string, respectively). We claimed in the bytecode spec that a fatal error would be thrown in this case, but this was not true. Instead, we'd just fail an assertion. This change should properly error in the way described in the spec.

Reviewed By: ricklavoie

Differential Revision: D5390092

fbshipit-source-id: 81b0aa32ddc277d60dc377b148f07963ebb8d259