fix runaway handlerstacks
Summary:
truncate handler stack when we hit a catch handler in an unwind (just like we already did for throw), which is the right thing to do. This
kills off some observed runaway handlerstacks. Also changed code to dump state in the (hopefully now impossible) case of a runaway, rather than just
failing with an exception.

Reviewed By: hubyrod

Differential Revision: D5397513

fbshipit-source-id: 7aa5d8af209c3d64e45adf46b63149ccbfe55dd5