Manually mark closures as unique in HHBBC index for non-whole-program mode
Summary:
When running HHBBC in non-whole-program mode, the systemlib Closure class may not be resolvable. As a consequence, the current Index::Index will not mark user defined closures as unique. This behavior will cause JIT to panic when dealing with CreateCl.

This diff patched Index::Index to manually mark closures as unique.

Reviewed By: markw65

Differential Revision: D5389104

fbshipit-source-id: 5ac716ac38e705fd87297c742bd77f8ba54705d8