Verifier checks that arguments to DefCls/DefFunc/DefTypeAlias exist
Summary: Opcodes referencing a class id, type alias id or function id were causing HHVM to crash if no function/type alias/function with that id existed in the unit. This adds a verifier check that the object referenced by the id exists. It also additionally verifies that the DefFunc opcode never has 0 as an immediate, as this refers to main and thus crashes HHVM.

Reviewed By: ricklavoie

Differential Revision: D5394665

fbshipit-source-id: 00ddbbe2fef52dd08ad5ebdff453b7ca1f52a312