Fix quadratic-time type check issue with generic parameter constraints
Summary:
For a function such as
```
function foo<T as C>(T $x):void
```
any calls to `foo` generate constraints on the fresh type variable used in place of `T`. These
constraints are not checked at the callsite, but are placed instead in a to-do list using `Env.add_todo`.

The constraints are checked at the end of type-checking the enclosing method or function body. But also: at every function call in
the body, for reasons not completely apparent to me. This means that as the list of todo's grows, it is repeatedly checked, with
consequent O(n^2) behaviour.

The solution for now is to remove to-do constraints as they are checked. Unfortunately this exposes a couple of latent type soundness
bugs in two tests. But problems with contravariance and unresolved types are known, and we're not really making this any worse
with this fix.

Also in this diff are some improvements to logging.

Reviewed By: dlreeves

Differential Revision: D5406520

fbshipit-source-id: c847ba0988fca2fdfd35ac0b1e147cc575348c14