Save a mov per iteration when probing hashes.
Summary:
Instead of keeping the full size hash around and making a masked copy every time we use it, we can keep just the smaller, already masked copy. This saves a mov per iteration in the probing loops.
Upgrading some variables from 32 to 64 bits was need to convince gcc that it doesn't need to insert movs to zero extend them.
Casting the hash to unsigned before masking is needed to convince gcc that it doesn't need to insert a movslq to sign extend it.

Reviewed By: mofarrell, edwinsmith

Differential Revision: D5403903

fbshipit-source-id: c41c188c211e254bc18d3df379d978615aa57826