Enforce that Map in HHBBC cannot represent packed arrays
Summary:
It makes the subtype, couldBe, and equals logic easier if Map cannot represent
packed arrays. This means that Packed and PackedN are always disjoint from
Map. This means that conversions to Map from static arrays can now fail, so
convert to Packed instead in that case.

MapN can still represent packed arrays. The reason for this is that we'll often
not initially know an array's keys, so we'll model it as a MapN, but then upon
later refinement, we'll infer the keys and realize its packed. If MapN is
disjoint from Packed and PackedN, it means the refinement isn't a subtype.

MapN already has a restriction that it cannot have a key type which is a
constant. A constant key implies it only represents arrays of size 1, which
means its not actually a superset of Map/Packed/PackedN. To simplify the logic,
this is forbidden. Add actual checking when creating a MapN for this case
instead of simply asserting it. Again, if the MapN conversion fails, then
perform a Map conversion instead.

Add more rigorous checking in checkInvariants(). Remove a few checkInvariant()
calls in subtypeOf() and the == operator. checkInvariants() calls subtypeOf()
and the == operator, and if a specialized array type has a lot of nested types,
we can get an exponential number of checkInvariant() calls. We call
checkInvariants() in many many places, so there's not much lost by omitting the
checks for these two operations.

Reviewed By: markw65

Differential Revision: D5405134

fbshipit-source-id: 4eb54486ccbf8b3d7e9d8c59e028d9bad11cfcfe