Stop eating the semicolon after a missing name in an expression
Summary:
In this diff, I improved error recover on a highly specific test case: an expression statement with the second of its variables omitted. Previously, the parser would generate two redundant errors on this mistake, but now it only generates the one that is relevant to the user.

Note that this particular check only fixes the case in which the parser encounters a token that is immediately needed by the parser one level up. So, this fix doesn't yet have any dramatic effects on reducing the depth of a cascading errors--it just removes duplicate errors in this one case.

Differential Revision: D5373930

fbshipit-source-id: c5911bdaf9793ed93427e6b7c022415c8803e025