Improving method error recovery, and adding scope to parser context
Summary:
**Overview:**

When inside a class, encountering a 'public,' 'private,' or 'protected' token inside an incomplete if_statement or methodish_declaration now bails out of the latter two scopes and continues parsing the-line-starting-with-the-visibility-modifier as a new methodish_declaration (rather than as a nested function, or as the body of an if statement). Testing shows that this greatly reduces the number of cascading errors.

Also, the parser context has been refactored to preserve a notion of "scope." Rather than storing context as a flat list of expected token_kinds, the token_kinds are bundled according to when in the context they became expected.

**More detail:**

This diff is another case-driven improvement to the full fidelity parser's error recovery. Specifically, it was motivated by the previously questionable recovery from the-programmer-hasn't-finished-typing-yet situations like those in test_nested_incomplete_if.php. It turns out that this file displayed two separate failures of parser error recovery, which I fixed.

//Previous behavior://
* The parser would correctly identify the if statement as missing its right paren, but would then treat the declaration of function g() as part of the body of the if.
* Even without the if statement, error recovery was weird--see test_nested_function(.php/.old). A new line starting with a visibility modifier would be parsed as a function nested inside of method g().
* Previous outputs: P57620004 for test_nested_if_incomplete.php, P57620022 for test_nested_function.php, and "(8,2)-(9,1) A right brace is expected here." for test_nested_function_lite.php.

//Current behavior://
* The parser knows that 'public,' 'private', and 'protected' are valid tokens to encounter inside a class. So, when it's inside a class, but specifically in a scope* where those tokens are unexpected--like in this incomplete if statement--it marks the rest of the if statement as 'missing' and bails out back up to the class level to continue parsing from there. (* well, not every such scope. Right now this just happens in if statements and methodish declarations, but this behavior can be expanded later.)
* When the parser is inside a method and encounters a 'public,' 'private,' or 'protected,' it marks the rest of the current method as 'missing' and starts parsing a new method at the visibility modifier.

Also, I refactored the parser context to include scope. I initially thought doing so would be necessary to solve the above issues, but that turned out to be untrue--my fixes don't make use of the scope feature. Still, I'm keeping the scope version around because it adds to the context extra information which might be useful later.

//Notes://
- HHVM forbids functions-inside-functions from having visibility modifiers, but previous FFP behavior on test_nested_incomplete_if.php shows that the FFP let them slide. This diff has the side effect of fixing that, though\!
- PHP allows nested nominal functions, but Hack doesn't. TODO introduce an error for that later.
- The .exp files still contain multiple errors at the same location because I'm running --full-fidelity-errors-all. HOWEVER, the full fidelity unit tests compare the .exp files to the output of --full-fidelity-errors, so I've chosen to exclude test_nested_incomplete_if from the unit tests.
- The presence and absence of comments appears to affect the location of error reporting. For example, moving the line-38 comment in test_nested_function.php to line 37, after the brace, changes the reported missing-brace location from (37,32)-(38,1) to (38,1)-(38,1). TODO fix that.

Differential Revision: D5384153

fbshipit-source-id: 5c5e92a59fa1cf37f60b6fb2eabec9a932265fdf