Set up for Nuclide Integration
Summary:
Now that some of the completions are done, I want to integrate what we have with the IDE so that we can ensure that I'm doing things in a way that is compatible with the IDE and also so we can do some manual testing/debugging.

This is not the full integration, but just some set up to do the full integration.

Reviewed By: ljw1004

Differential Revision: D5335083

fbshipit-source-id: 70e4ef862abd6a05cf56488c40236aec4ca944bc